requirements.txt
Pillow==9.5.0
torch==2.3.1+cu121
transformers==4.46.3
numpy<2
einops
timm
